feat: allowed FE features in browser
