+++ api/fe.js
@@ -1,0 +1,5 @@
+const { generateScript } = require('../forward_engineering/api');
+
+module.exports = {
+	generateScript,
+};
