+++ shared/readConfig.js
@@ -1,0 +1,15 @@
+const fs = require('fs');
+const path = require('path');
+
+const readConfig = pathToConfig => {
+	return JSON.parse(
+		fs
+			.readFileSync(path.join(__dirname, pathToConfig))
+			.toString()
+			.replace(/\/\*[.\s]*?\*\//gi, ''),
+	);
+};
+
+module.exports = {
+	readConfig,
+};
